requirements.txt
numpy
Flask
-e .
gunicorn
sklearn
pandas
PyYAML

